Revert "refactor: activity id Integer로 변경"
This reverts commit bea44bb4c28c7e26b7494fcface5f8e95269f9cb.

+++ src/main/java/in/koreatech/koin/domain/activity/dto/ActivityResponse.java
@@ -38,12 +38,11 @@ public record ActivityResponse(
     List<String> imageUrls,
 
     @Schema(description = "고유 식별자", example = "1")
-    Integer id,
+    Long id,
 
     @Schema(description = "제목", example = "코인 시간표 기능 추가")
     String title
 ) {
-
     public static ActivityResponse of(Activity activity, List<String> imageUrls) {
         return new ActivityResponse(
             activity.getDate(),
